[refactor] 파일 이동 및 스타일 수정
