+++ src/routes/setting/SettingRoutes.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import {createNativeStackNavigator} from '@react-navigation/native-stack';
 import AlarmSettingScreen from '~/screens/setting/updateAlarm/AlarmSettingScreen';
-import EditProfileScreen from '~/screens/setting/updateProfile/EditProfileScreen';
+import EditProfileScreen from '~/screens/setting/EditProfileScreen';
 import LeaveScreen from '~/screens/setting/LeaveScreen';
 import {SettingStackParamList} from '~/utils/types';
 import FavoriteListScreen from '~/screens/setting/favoriteList/FavoriteListScreen';
